requirements.txt
fastrtc[vad]>=0.0.19
groq>=0.22.0
numpy>=2.1.3
langgraph>=0.1.18
langchain-core>=0.1.29
langchain-groq>=0.1.5
loguru>=0.7.3
python-dotenv
tavily-python
soundfile
librosa
langchain-tavily
cartesia
gradio==5.29.1
gradio-client==1.10.1
google-genai>=0.2.0
websockets>=12.0
Pillow>=10.0.0
langchain-community
langchain-cerebras
langchain-google-genai
yfinance
pyowm
firecrawl-py
exa-py
kokoro-onnx>=0.4.7